Move extension method classes to a separate folder
