+++ src/NLog.Targets.Syslog/LogEventMsgSet.cs
@@ -4,6 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using NLog.Common;
 using NLog.Layouts;
+using NLog.Targets.Syslog.Extensions;
 using NLog.Targets.Syslog.MessageCreation;
 using NLog.Targets.Syslog.MessageSend;
