+++ src/NLog.Targets.Syslog/MessageCreation/SdParam.cs
@@ -3,6 +3,7 @@ using NLog.Layouts;
 using NLog.Targets.Syslog.Policies;
 using System.Collections.Generic;
 using System.Linq;
+using NLog.Targets.Syslog.Extensions;
 
 namespace NLog.Targets.Syslog.MessageCreation
 {
