+++ src/NLog.Targets.Syslog/Policies/Unidecoder.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using NLog.Targets.Syslog.Extensions;
 
 namespace NLog.Targets.Syslog.Policies
 {
